Feat/make chat list functionable

+++ composeApp/src/commonMain/kotlin/io/adik5050/discord_like/ui/app/navigation/main/MainNavigation.kt
@@ -31,7 +31,7 @@ fun MainNavigation(
     appDatabase: AppDatabase,
     userSession: UserSession,
     windowSizeClass: WindowSizeClass,
-    onNavigateToChat: () -> Unit,
+    onNavigateToChat: (Int) -> Unit,
     onNavigateToEditProfile: () -> Unit,
     onNavigateToWelcome: () -> Unit,
     onNavigateToCreateChannel: () -> Unit,
